# openapi.JSON
{
  "openapi": "3.0.0",
  "info": {
    "title": "Task API",
    "version": "1.0.0",
    "description": "A simple CRUD API managing tasks"
  },
  "paths": {
    "/": {
      "get": {
        "summary": "API Root",
        "description": "Returns API metadata",
        "responses": {
          "200": { "description": "Successful response" }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health Check",
        "description": "Checks if the server is running",
        "responses": {
          "200": { "description": "Server is alive" }
        }
      }
    },
    "/tasks": {
      "get": {
        "summary": "Get all tasks",
        "description": "Retrieves the complete list of tasks",
        "responses": {
          "200": { "description": "List of tasks returned successfully" }
        }
      },
      "post": {
        "summary": "Create a task",
        "description": "Adds a new task to the list",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" }
                },
                "required": ["title"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Task created successfully" },
          "400": { "description": "Bad Request - Missing or empty title" }
        }
      }
    },
    "/tasks/{id}": {
      "get": {
        "summary": "Get a single task",
        "description": "Retrieves a specific task by its ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Task found" },
          "404": { "description": "Task not found" }
        }
      },
      "put": {
        "summary": "Update a task",
        "description": "Updates an existing task's title or done status",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "done": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Task updated successfully" },
          "400": { "description": "Invalid input/body" },
          "404": { "description": "Task not found" }
        }
      },
      "delete": {
        "summary": "Delete a task",
        "description": "Removes a task by its ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "204": { "description": "Task deleted successfully" },
          "404": { "description": "Task not found" }
        }
      }
    }
  }
}